% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrons.R
\docType{data}
\name{electrons}
\alias{electrons}
\title{Fifteen determinations of the charge of an electron.  This is Wheeler's first example of the use of the W-Ratio in his article.}
\source{
\url{https://www.qualitydigest.com/inside/quality-insider-article/analysis-using-few-data-part-two.html}
}
\description{
Fifteen determinations of the charge of an electron.  This is Wheeler's first example of the use of the W-Ratio in his article.
}
\references{
Wheeler, Donald J 2012, Analysis Using Few Data Parts 1 and 2, Quality Digest

\url{https://www.nobelprize.org/prizes/physics/1923/millikan/biographical/}
}
\keyword{data}
