% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCrit.r
\name{WCrit}
\alias{WCrit}
\title{W-Ratio Critical Values}
\usage{
WCrit(size, alpha)
}
\arguments{
\item{size}{Length of Data Set.  Must between 3 and 20.}

\item{alpha}{Desired Alpha Level. Must be 0.05, 0.10, 0.15, or 0.2.}
}
\description{
This function returns the critical values for the W-Ratio Test.
}
\examples{
WCrit(3,.01)
}
