% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRatios.r
\name{WRatios}
\alias{WRatios}
\title{W-Ratio Values}
\usage{
WRatios(data)
}
\arguments{
\item{data}{Dataset must be a vector.}
}
\description{
This function returns the W-Ratios of a dataset.
}
\examples{
WRatios(electrons)
}
