% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRatioTest.R
\name{WRatioTest}
\alias{WRatioTest}
\title{W-Ratio Test}
\usage{
WRatioTest(data, alpha = NULL)
}
\arguments{
\item{data}{Dataset.  Must be between 3 and 20 values.}

\item{alpha}{Desired Alpha Level. Must be 0.05, 0.10, 0.15, 0.2, or NULL.  If NULL is selected alpha will be set based upon the dataset size.}
}
\description{
This function performs the W-Ratio test on a data set.  This also returns a data frame with the detailed test results.
}
\examples{
WRatioTest(electrons)
}
